Extract multiview video player to its own compoent
